Move peaklet processing to separate file, remove old peaks
