+++ straxen/plugins/__init__.py
@@ -1,14 +1,14 @@
 from . import daqreader
 from .daqreader import *
 
-from . import peak_processing
-from .peak_processing import *
-
 from . import pax_interface
 from .pax_interface import *
 
 from . import pulse_processing
 from .pulse_processing import *
+
+from . import peaklet_processing
+from .peaklet_processing import *
 
 from . import peak_processing
 from .peak_processing import *
